Fix a variety of type mismatches reported by splint.
git-svn-id: svn://svn.code.sf.net/p/bogofilter/code/trunk/bogofilter@3462 046ba1d7-8659-4f4f-8137-08d775d845e8

+++ src/bogoreader.c
@@ -162,7 +162,7 @@ static st_t ismaildir(const char *dir) {
     size_t l;
     char *x;
     const char *const *y;
-    const int maxlen = 4;
+    const size_t maxlen = 4;
 
     r = isdir(dir);
     if (r != IS_DIR) return r;
